Move rawhid to lib
